+++ Changelog.md
@@ -8,6 +8,10 @@
   - test: Adding missing tests
   - chore: Changes to the build process or auxiliary tools/libraries/documentation
 
+## 0.9.0
+- refactor: rename `Excel` in `Sheets` to accomodate OpenDocuments
+- feat: add Index/IndexMut for Range
+
 ## 0.8.0
 - feat: add basic support for opendocument spreadsheets
 - style: apply rustfmt